fix(sso)!: unify SAML response processing, fix provider lookup and config bugs
Extract duplicated SAML response handling from callbackSSOSAML and
acsEndpoint into a shared processSAMLResponse pipeline. Complete
createSP/createIdP helpers with all encryption and signing fields so
both endpoints produce identical SP/IdP objects.

Fixes:
- SP metadata endpoint used internal row ID instead of public providerId
  in ACS URL fallback
- acsEndpoint skipped DB lookup when defaultSSO was configured
- acsEndpoint fell back to defaultSSO[0] for unmatched providerIds
- acsEndpoint was missing encryption fields (isAssertionEncrypted,
  encPrivateKey) causing silent decryption failures
- Registration accepted broken configs with empty entryPoint and no
  IdP metadata
- Supabase migration guide generated unsupported metadataUrl field
- SSO docs had contradictory callbackUrl guidance
